+++ src/components/atoms/vRoomAvatar.unit.ts
@@ -12,6 +12,14 @@ const mockData = {
 	yPosition: 2,
 };
 
+const propsData = {
+	item: mockData,
+	size: 100,
+	groupAvatar: false,
+	showBadge: true,
+	location: { x: 5, y: 2 },
+};
+
 describe("vRoomAvatar", () => {
 	it("should display the title", () => {
 		const wrapper = mount(vRoomAvatar, {
@@ -19,12 +27,7 @@ describe("vRoomAvatar", () => {
 				i18n: true,
 				vuetify: true,
 			}),
-			propsData: {
-				item: mockData,
-				size: 100,
-				groupAvatar: false,
-				showBadge: true,
-			},
+			propsData,
 		});
 		const labelElement = wrapper.find(".sub-title").element as HTMLElement;
 		expect(labelElement).toBeTruthy();
@@ -37,12 +40,7 @@ describe("vRoomAvatar", () => {
 				i18n: true,
 				vuetify: true,
 			}),
-			propsData: {
-				item: mockData,
-				size: 100,
-				groupAvatar: false,
-				showBadge: true,
-			},
+			propsData,
 		});
 		const shortLabelElement = wrapper.find(".single-avatar")
 			.element as HTMLElement;
@@ -61,6 +59,7 @@ describe("vRoomAvatar", () => {
 				size: 100,
 				groupAvatar: false,
 				showBadge: true,
+				location: { x: 5, y: 2 },
 			},
 		});
 		const badgeElement = wrapper.find(".badge-component");
@@ -75,12 +74,7 @@ describe("vRoomAvatar", () => {
 				i18n: true,
 				vuetify: true,
 			}),
-			propsData: {
-				item: mockData,
-				size: 100,
-				groupAvatar: false,
-				showBadge: true,
-			},
+			propsData,
 		});
 		const badgeElement = wrapper.find(".badge-component");
 		expect(badgeElement).toBeTruthy();
@@ -88,23 +82,20 @@ describe("vRoomAvatar", () => {
 		expect(badgeElement.vm.$data.isActive).toBeFalsy();
 	});
 
-	it("should display the correct color and size", () => {
+	it("should display the correct color and size", async () => {
 		const wrapper = mount(vRoomAvatar, {
 			...createComponentMocks({
 				i18n: true,
 				vuetify: true,
 			}),
-			propsData: {
-				item: mockData,
-				size: 100,
-				groupAvatar: false,
-				showBadge: true,
-			},
+			propsData,
 		});
 		const avatarComponent = wrapper.find(".avatar-component");
 		expect(avatarComponent).toBeTruthy();
 		expect(avatarComponent.vm.$props.color).toStrictEqual("#ffffff");
 		expect(avatarComponent.vm.$props.size).toStrictEqual(100);
+		await wrapper.setProps({ isGroupAvatarList: true });
+		expect(avatarComponent.vm.$props.size).toStrictEqual(75);
 	});
 
 	it("should have 'rounded-xl' class name", () => {
@@ -113,12 +104,7 @@ describe("vRoomAvatar", () => {
 				i18n: true,
 				vuetify: true,
 			}),
-			propsData: {
-				item: mockData,
-				size: 100,
-				groupAvatar: false,
-				showBadge: true,
-			},
+			propsData,
 		});
 		const avatarComponent = wrapper.find(".avatar-component");
 		expect(avatarComponent).toBeTruthy();
@@ -136,6 +122,7 @@ describe("vRoomAvatar", () => {
 				size: 100,
 				groupAvatar: true,
 				showBadge: true,
+				location: { x: 5, y: 2 },
 			},
 		});
 		const avatarComponent = wrapper.find(".avatar-component");
@@ -149,12 +136,7 @@ describe("vRoomAvatar", () => {
 				i18n: true,
 				vuetify: true,
 			}),
-			propsData: {
-				item: mockData,
-				size: 100,
-				groupAvatar: false,
-				showBadge: true,
-			},
+			propsData,
 		});
 		const avatarComponent = wrapper.find(".avatar-component");
 
@@ -164,5 +146,99 @@ describe("vRoomAvatar", () => {
 		let emitted = wrapper.emitted();
 		expect(emitted["click"]).toHaveLength(1);
 		expect(emitted["click"] && emitted["click"][0][0]).toStrictEqual(mockData);
+	});
+
+	it("should emit 'dragStart' event when it started dragging", async () => {
+		const wrapper = mount(vRoomAvatar, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+			}),
+			propsData,
+		});
+		const avatarComponent = wrapper.find(".avatar-component");
+		expect(wrapper.vm.$data.isDragging).toBe(false);
+
+		avatarComponent.trigger("dragstart");
+		await wrapper.vm.$nextTick();
+
+		expect(wrapper.vm.$data.isDragging).toBe(true);
+
+		let emitted = wrapper.emitted();
+		expect(emitted["startDrag"]).toHaveLength(1);
+		expect(emitted["startDrag"] && emitted["startDrag"][0][0]).toStrictEqual(
+			mockData
+		);
+		expect(emitted["startDrag"] && emitted["startDrag"][0][1]).toStrictEqual({
+			x: 5,
+			y: 2,
+		});
+	});
+
+	it("should emit 'drop' event when an element dropped onto it", async () => {
+		const wrapper = mount(vRoomAvatar, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+			}),
+			propsData,
+		});
+		const avatarComponent = wrapper.find(".avatar-component");
+
+		avatarComponent.trigger("drop");
+		await wrapper.vm.$nextTick();
+
+		let emitted = wrapper.emitted();
+		expect(emitted["drop"]).toHaveLength(1);
+		expect(emitted["drop"] && emitted["drop"][0][0]).toStrictEqual({
+			x: 5,
+			y: 2,
+		});
+	});
+
+	it("should change its size while dragging", async () => {
+		const wrapper = mount(vRoomAvatar, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+			}),
+			propsData,
+		});
+		await wrapper.setProps({ isGroupAvatarList: true });
+		const avatarComponent = wrapper.find(".avatar-component");
+		expect(wrapper.vm.$data.isDragging).toBe(false);
+		expect(avatarComponent.vm.$props.size).toStrictEqual(75);
+
+		avatarComponent.trigger("dragstart");
+		await wrapper.vm.$nextTick();
+
+		expect(wrapper.vm.$data.isDragging).toBe(true);
+
+		expect(avatarComponent.vm.$props.size).toStrictEqual(100);
+	});
+
+	it("should emit payload as '{x:-1, y:-1}' if the 'isGroupAvatarList' prop is set true while dragging", async () => {
+		const wrapper = mount(vRoomAvatar, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+			}),
+			propsData,
+		});
+		await wrapper.setProps({ isGroupAvatarList: true });
+		const avatarComponent = wrapper.find(".avatar-component");
+
+		avatarComponent.trigger("dragstart");
+		await wrapper.vm.$nextTick();
+
+		let emitted = wrapper.emitted();
+		expect(emitted["startDrag"]).toHaveLength(1);
+		expect(emitted["startDrag"] && emitted["startDrag"][0][0]).toStrictEqual(
+			mockData
+		);
+		expect(emitted["startDrag"] && emitted["startDrag"][0][1]).toStrictEqual({
+			x: -1,
+			y: -1,
+		});
 	});
 });