+++ src/components/templates/TasksDashboardTeacher.unit.js
@@ -1,5 +1,6 @@
 import TasksDashboardTeacher from "./TasksDashboardTeacher";
 import TasksList from "@components/organisms/TasksList";
+import vCustomEmptyState from "@components/molecules/vCustomEmptyState";
 import {
 	overDueTasksTeacher,
 	dueDateTasksTeacher,
@@ -13,11 +14,15 @@ describe("@components/templates/TasksDashboardTeacher", () => {
 			getters: {
 				getStatus: () => "completed",
 				hasNoTasks: () => false,
+				hasNoOpenTasksTeacher: () => false,
+				hasNoDrafts: () => true,
 				getOpenTasksForTeacher: () => ({
 					overdue: overDueTasksTeacher,
 					withDueDate: dueDateTasksTeacher,
 					noDueDate: noDueDateTasksTeacher,
 				}),
+				getDraftTasksForTeacher: () => [],
+				hasFilterSelected: () => false,
 			},
 		},
 	};
@@ -38,6 +43,9 @@ describe("@components/templates/TasksDashboardTeacher", () => {
 				store: mockStore,
 			}),
 			vuetify,
+			propsData: {
+				tab: 0,
+			},
 		});
 
 		const expansionPanels = wrapper.findAll(".v-expansion-panel");
@@ -50,5 +58,57 @@ describe("@components/templates/TasksDashboardTeacher", () => {
 		expect(expansionPanels.at(1).classes()).toContain(
 			"v-expansion-panel--active"
 		);
+	});
+
+	it("Should render empty state when there are no drafts", () => {
+		const wrapper = mount(TasksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStore,
+			}),
+			vuetify,
+			propsData: {
+				tab: 1,
+			},
+		});
+
+		expect(wrapper.findComponent(vCustomEmptyState).exists()).toBe(true);
+	});
+
+	it("Should render empty state when there are no drafts due to filter", () => {
+		const wrapper = mount(TasksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStore,
+			}),
+			vuetify,
+			propsData: {
+				tab: 1,
+				hasFilterSelected: true,
+			},
+		});
+
+		expect(wrapper.findComponent(vCustomEmptyState).exists()).toBe(true);
+	});
+
+	it("Should trigger event to update tab property", async () => {
+		const wrapper = mount(TasksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStore,
+			}),
+			vuetify,
+			propsData: {
+				tab: 1,
+			},
+		});
+
+		await wrapper.setData({ currentTab: 0 });
+
+		expect(wrapper.emitted("update:tab")).toHaveLength(1);
+		expect(wrapper.emitted("update:tab")[0]).toStrictEqual([0]);
 	});
 });