+++ src/components/templates/TasksDashboardTeacher.vue
@@ -1,37 +1,50 @@
 <template>
 	<section class="task-dashboard-teacher">
-		<v-custom-double-panels
-			:panel-one-count="noDueDateTasks.length"
-			:panel-two-count="withDueDateTasks.length + overdueTasks.length"
-			:panel-one-title="$t('pages.tasks.subtitleNoDue')"
-			:panel-two-title="$t('pages.tasks.subtitleWithDue')"
-			:status="status"
-			:is-empty="hasNoTasks"
-			:expanded-default="1"
-		>
-			<template v-slot:panelOne>
-				<tasks-list :tasks="noDueDateTasks" type="teacher" />
-			</template>
-			<template v-slot:panelTwo>
-				<tasks-list
-					:tasks="overdueTasks"
-					:title="$t('pages.tasks.teacher.subtitleOverDue')"
-					type="teacher"
+		<v-tabs-items v-model="currentTab">
+			<v-tab-item>
+				<v-custom-double-panels
+					:panel-one-count="noDueDateTasks.length"
+					:panel-two-count="withDueDateTasks.length + overdueTasks.length"
+					:panel-one-title="$t('pages.tasks.subtitleNoDue')"
+					:panel-two-title="$t('pages.tasks.subtitleWithDue')"
+					:status="status"
+					:is-empty="hasNoOpenTasksTeacher"
+					:expanded-default="1"
+				>
+					<template v-slot:panelOne>
+						<tasks-list :tasks="noDueDateTasks" type="teacher" />
+					</template>
+					<template v-slot:panelTwo>
+						<tasks-list
+							:tasks="overdueTasks"
+							:title="$t('pages.tasks.teacher.subtitleOverDue')"
+							type="teacher"
+						/>
+						<tasks-list
+							:tasks="withDueDateTasks"
+							:title="$t('pages.tasks.subtitleOpen')"
+							type="teacher"
+						/>
+					</template>
+				</v-custom-double-panels>
+				<v-custom-empty-state
+					v-if="hasNoOpenTasksTeacher"
+					:image="emptyStateImage"
+					:title="emptyStateText.title"
+					:subtitle="emptyStateText.subtitle"
+					class="mt-16"
 				/>
-				<tasks-list
-					:tasks="withDueDateTasks"
-					:title="$t('pages.tasks.subtitleOpen')"
-					type="teacher"
+			</v-tab-item>
+			<v-tab-item>
+				<tasks-list :tasks="draftTasks" type="teacher" />
+				<v-custom-empty-state
+					v-if="hasNoDrafts"
+					:image="emptyStateImage"
+					:title="emptyStateText.title"
+					class="mt-16"
 				/>
-			</template>
-		</v-custom-double-panels>
-		<v-custom-empty-state
-			v-if="hasNoTasks"
-			:image="emptyStateImage"
-			:title="$t('pages.tasks.teacher.emptyState.title')"
-			:subtitle="$t('pages.tasks.teacher.emptyState.subtitle')"
-			class="mt-16"
-		/>
+			</v-tab-item>
+		</v-tabs-items>
 	</section>
 </template>
 
@@ -44,6 +57,12 @@ import tasksEmptyState from "@assets/img/empty-state/Task_Empty_State.svg";
 
 export default {
 	components: { vCustomEmptyState, TasksList, vCustomDoublePanels },
+	props: {
+		tab: {
+			type: Number,
+			required: true,
+		},
+	},
 	data() {
 		return {
 			emptyStateImage: tasksEmptyState,
@@ -52,8 +71,12 @@ export default {
 	computed: {
 		...mapGetters("tasks", {
 			openTasks: "getOpenTasksForTeacher",
+			draftTasks: "getDraftTasksForTeacher",
 			status: "getStatus",
 			hasNoTasks: "hasNoTasks",
+			hasNoOpenTasksTeacher: "hasNoOpenTasksTeacher",
+			hasNoDrafts: "hasNoDrafts",
+			hasFilterSelected: "hasFilterSelected",
 		}),
 		overdueTasks: function () {
 			return this.openTasks.overdue;
@@ -63,6 +86,34 @@ export default {
 		},
 		withDueDateTasks: function () {
 			return this.openTasks.withDueDate;
+		},
+		emptyStateText: function () {
+			if (this.hasFilterSelected) {
+				return {
+					title: this.$t("pages.tasks.emptyStateOnFilter.title"),
+					subtitle: undefined,
+				};
+			} else {
+				if (this.tab === 0) {
+					return {
+						title: this.$t("pages.tasks.teacher.emptyState.title"),
+						subtitle: this.$t("pages.tasks.teacher.emptyState.subtitle"),
+					};
+				} else {
+					return {
+						title: this.$t("pages.tasks.teacher.drafts.emptyState.title"),
+						subtitle: undefined,
+					};
+				}
+			}
+		},
+		currentTab: {
+			get() {
+				return this.tab;
+			},
+			set(newTab) {
+				this.$emit("update:tab", newTab);
+			},
 		},
 	},
 };